+++ scripts/jira-create-version-pipelines-ui.js
@@ -75,12 +75,14 @@ function getVersionIssues(versionId) {
         return reject(error);
       } else {
         const search = stdout.match(versionId);
-        const tmp = stdout.substring(0, search.index).match(/MS-\d{4,}/g);
+        const tmp = stdout.substring(0, search.index).match(/MS-\d{4,}/g) || [];
 
         let issues = [];
-        for (var i = 0; i < tmp.length; i++) {
-          if (issues.indexOf(tmp[i]) == -1) {
-            issues.push(tmp[i]);
+        if (tmp != null) {
+          for (var i = 0; i < tmp.length; i++) {
+            if (issues.indexOf(tmp[i]) == -1) {
+              issues.push(tmp[i]);
+            }
           }
         }
         console.log(issues);